Move RDBMS class tests to their sub-directories
